backend/requirements.txt
fastapi[standard]
sqlalchemy[asyncio]>=2.0
aiosqlite
faster-whisper
ollama
python-pptx
uvicorn[standard]
python-multipart
av
stripe
openai
python-dotenv
